Windows plugin working, 'update' command still needs elevated privileges but all other commands can be run as normal user.

+++ VCamUtils/src/fraction.cpp
@@ -18,7 +18,6 @@
  */
 
 #include <cstdlib>
-#include <cwchar>
 #include <sstream>
 #include <string>
 
@@ -133,7 +132,7 @@ int64_t &AkVCam::Fraction::den()
 
 double AkVCam::Fraction::value() const
 {
-    return double(this->d->m_num) / this->d->m_den;
+    return double(this->d->m_num) / double(this->d->m_den);
 }
 
 std::string AkVCam::Fraction::toString() const
